restructured project to handle routes
