+++ package.json
@@ -8,6 +8,7 @@
     "jquery": "^3.3.1",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "^2.1.3",
     "save-dev": "^2.0.0"
   },
