+++ src/components/Recipes.jsx
@@ -8,7 +8,8 @@ export class Recipes extends Component {
 
         if (this.props.isLoading) {
             return <p>Loading ...</p>;
-        } else if (this.props.recipes.length === 0) {
+        }
+        else if (this.props.recipes.length === 0) {
             return <h1>Sorry, no search results!</h1>;
         }
         return (
